requirements.txt
pybullet
gym
pyyaml
torch
torchvision
baselines
stable-baselines
numpy 
seaborn
imageio
hydra-core
sklearn
umap-learn
umap-learn[plot]
pandas
matplotlib
datashader
bokeh
holoviews
colorcet
gast==0.2.2
